requirements.txt
streamlit==1.40.2
Pillow==10.3.0
torch==2.4.1
torchvision==0.19.1